Feat#52 - Smtp 회원가입 인증 적용

+++ neulbomgil-backend/src/main/java/com/gooroomees/neulbomgil_backend/domain/auth/service/UserAuthService.java
@@ -10,8 +10,8 @@ import org.springframework.stereotype.Service;
 public class UserAuthService {
     private final UserAuthRepository userAuthRepository;
 
-    public UserAuth findById(int userId) {
-        return userAuthRepository.findById(userId)
+    public UserAuth findById(Long userId) {
+        return userAuthRepository.findById(userId.intValue())
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 사용자입니다."));
     }
 }
